ref(numberManager): add numberManager pakage to split class.
