+++ src/main/java/calculator/Printer.java
@@ -4,7 +4,8 @@ public class Printer {
 
     static int first = 0, second = 0, third = 0, fourth = 0, fifth = 0;
     static int totalIncome = 0;
-    public static void printResult(){
+
+    public void printResult(){
         System.out.println("당첨 통계\n" +
                 "---\n" +
                 "3개 일치 (5,000원) - " + fifth + "개\n" +
@@ -14,7 +15,7 @@ public class Printer {
                 "6개 일치 (2,000,000,000원) - " + first + "개");
     }
 
-    public static void printIncomeRate(int coin){
+    public void printIncomeRate(int coin){
         double incomeRateTemp = ((double)totalIncome / (double)(coin * 1000))*100;
         double incomeRate = Math.round(incomeRateTemp*100)/100.0;
         System.out.println("총 수익률은 " + incomeRate + "%입니다.");
